SelfTests: Moved cCompositeChat test to a separate project.
