+++ src/CompositeChat.cpp
@@ -6,95 +6,6 @@
 #include "Globals.h"
 #include "CompositeChat.h"
 #include "ClientHandle.h"
-#include "SelfTests.h"
-
-
-
-
-
-#ifdef SELF_TEST
-
-/** A simple self-test that verifies that the composite chat parser is working properly. */
-class SelfTest_CompositeChat
-{
-public:
-	SelfTest_CompositeChat(void)
-	{
-		cSelfTests::Get().Register(cSelfTests::SelfTestFunction(&TestParser1), "CompositeChat parser test 1");
-		cSelfTests::Get().Register(cSelfTests::SelfTestFunction(&TestParser2), "CompositeChat parser test 2");
-		cSelfTests::Get().Register(cSelfTests::SelfTestFunction(&TestParser3), "CompositeChat parser test 3");
-		cSelfTests::Get().Register(cSelfTests::SelfTestFunction(&TestParser4), "CompositeChat parser test 4");
-		cSelfTests::Get().Register(cSelfTests::SelfTestFunction(&TestParser5), "CompositeChat parser test 5");
-	}
-
-	static void TestParser1(void)
-	{
-		cCompositeChat Msg;
-		Msg.ParseText("Testing @2color codes and http://links parser");
-		const cCompositeChat::cParts & Parts = Msg.GetParts();
-		assert_test(Parts.size() == 4);
-		assert_test(Parts[0]->m_PartType == cCompositeChat::ptText);
-		assert_test(Parts[1]->m_PartType == cCompositeChat::ptText);
-		assert_test(Parts[2]->m_PartType == cCompositeChat::ptUrl);
-		assert_test(Parts[3]->m_PartType == cCompositeChat::ptText);
-		assert_test(Parts[0]->m_Style == "");
-		assert_test(Parts[1]->m_Style == "@2");
-		assert_test(Parts[2]->m_Style == "@2");
-		assert_test(Parts[3]->m_Style == "@2");
-	}
-
-	static void TestParser2(void)
-	{
-		cCompositeChat Msg;
-		Msg.ParseText("@3Advanced stuff: @5overriding color codes and http://links.with/@4color-in-them handling");
-		const cCompositeChat::cParts & Parts = Msg.GetParts();
-		assert_test(Parts.size() == 4);
-		assert_test(Parts[0]->m_PartType == cCompositeChat::ptText);
-		assert_test(Parts[1]->m_PartType == cCompositeChat::ptText);
-		assert_test(Parts[2]->m_PartType == cCompositeChat::ptUrl);
-		assert_test(Parts[3]->m_PartType == cCompositeChat::ptText);
-		assert_test(Parts[0]->m_Style == "@3");
-		assert_test(Parts[1]->m_Style == "@5");
-		assert_test(Parts[2]->m_Style == "@5");
-		assert_test(Parts[3]->m_Style == "@5");
-	}
-
-	static void TestParser3(void)
-	{
-		cCompositeChat Msg;
-		Msg.ParseText("http://links.starting the text");
-		const cCompositeChat::cParts & Parts = Msg.GetParts();
-		assert_test(Parts.size() == 2);
-		assert_test(Parts[0]->m_PartType == cCompositeChat::ptUrl);
-		assert_test(Parts[1]->m_PartType == cCompositeChat::ptText);
-		assert_test(Parts[0]->m_Style == "");
-		assert_test(Parts[1]->m_Style == "");
-	}
-
-	static void TestParser4(void)
-	{
-		cCompositeChat Msg;
-		Msg.ParseText("links finishing the text: http://some.server");
-		const cCompositeChat::cParts & Parts = Msg.GetParts();
-		assert_test(Parts.size() == 2);
-		assert_test(Parts[0]->m_PartType == cCompositeChat::ptText);
-		assert_test(Parts[1]->m_PartType == cCompositeChat::ptUrl);
-		assert_test(Parts[0]->m_Style == "");
-		assert_test(Parts[1]->m_Style == "");
-	}
-
-	static void TestParser5(void)
-	{
-		cCompositeChat Msg;
-		Msg.ParseText("http://only.links");
-		const cCompositeChat::cParts & Parts = Msg.GetParts();
-		assert_test(Parts.size() == 1);
-		assert_test(Parts[0]->m_PartType == cCompositeChat::ptUrl);
-		assert_test(Parts[0]->m_Style == "");
-	}
-
-} gTest;
-#endif  // SELF_TEST
