+++ tests/CompositeChat/CMakeLists.txt
@@ -1,0 +1,47 @@
+cmake_minimum_required (VERSION 2.6)
+
+enable_testing()
+add_definitions(-DTEST_GLOBALS=1)
+
+include_directories(${CMAKE_CURRENT_SOURCE_DIR})
+include_directories(${CMAKE_SOURCE_DIR}/src/)
+include_directories(${CMAKE_SOURCE_DIR}/lib/jsoncpp/include)
+
+add_definitions(-DTEST_GLOBALS=1)
+
+set (SHARED_SRCS
+	${CMAKE_SOURCE_DIR}/src/CompositeChat.cpp
+	${CMAKE_SOURCE_DIR}/src/OSSupport/StackTrace.cpp
+)
+
+set (SHARED_HDRS
+	${CMAKE_SOURCE_DIR}/src/CompositeChat.h
+	${CMAKE_SOURCE_DIR}/src/OSSupport/StackTrace.h
+)
+
+set (SRCS
+	CompositeChatTest.cpp
+	ClientHandle.cpp
+)
+
+if (MSVC)
+	# Add the MSVC-specific LeakFinder / StackTracer sources:
+	list (APPEND SHARED_SRCS ${CMAKE_SOURCE_DIR}/src/LeakFinder.cpp ${CMAKE_SOURCE_DIR}/src/StackWalker.cpp)
+	list (APPEND SHARED_HDRS ${CMAKE_SOURCE_DIR}/src/LeakFinder.h   ${CMAKE_SOURCE_DIR}/src/StackWalker.h)
+endif()
+
+source_group("Shared" FILES ${SHARED_SRCS} ${SHARED_HDRS})
+source_group("Sources" FILES ${SRCS})
+add_executable(CompositeChat-exe ${SRCS} ${SHARED_SRCS} ${SHARED_HDRS})
+target_link_libraries(CompositeChat-exe jsoncpp_lib_static)
+add_test(NAME CompositeChat-test COMMAND CompositeChat-exe)
+
+
+
+
+
+# Put the projects into solution folders (MSVC):
+set_target_properties(
+	CompositeChat-exe
+	PROPERTIES FOLDER Tests
+)
